refactor: create dedicated theme folder to handle custom MUI themes
