+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@emotion/react": "^11.11.3",
     "@emotion/styled": "^11.11.0",
+    "@material-ui/core": "^4.12.4",
     "@mui/material": "^5.15.4",
     "@next/font": "^14.0.4",
     "next": "14.0.4",
