+++ src/app/page.js
@@ -1,10 +1,13 @@
 import Image from 'next/image';
-import Landing from './layout/Landing';
+import Landing from '@/layouts/Landing';
+import TitleCard from '@/components/TitleCard';
 
 export default function Home() {
+  const titleText = "Dungeon As Bro"
+
   return (
     <Landing>
-      <div id="test">Dungeon As Bro</div>
+      <TitleCard titleText={titleText}/>
     </Landing>
   )
 };
